requirements.txt
argcomplete==1.12.1
boto3==1.15.12
botocore==1.18.12
certifi==2020.6.20
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
durationpy==0.5
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
hjson==3.0.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==1.1.1
numpy==1.19.2
opencv-contrib-python==4.4.0.44
pdfkit==0.6.1
pip-tools==5.3.1
placebo==0.9.0
protobuf==4.0.0rc2
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.6.1
python-decouple==3.3
python-slugify==4.0.1
PyYAML==5.3.1
pyzbar==0.1.8
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.19
stripe==2.54.0
text-unidecode==1.3
toml==0.10.1
tqdm==4.50.0
troposphere==2.6.2
urllib3==1.25.10
Werkzeug==0.16.1
wkhtmltopdf==0.2
wsgi-request-logger==0.4.6
zappa==0.51.0
